Permission - Creating Permissions for Project

+++ app/Http/Controllers/Admin/RoleUserController.php
@@ -112,6 +112,9 @@ class RoleUserController extends Controller
     public function destroy(string $id)
     {
         $user = Admin::findOrFail($id);
+        if($user->getRoleNames()->first() === 'Super Admin'){
+            return response(['status' => 'error', 'message' => __('Can\'t Delete the Super User')]);
+        }
         $user->delete();
 
         return response(['status' => 'success', 'message' => __('Deleted Successfully')]);
